backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
psycopg[binary]==3.1.18
pgvector==0.2.5
celery==5.3.6
redis==5.0.8
flasgger==0.9.7.1
python-json-logger==2.0.7
fpdf2==2.7.9
twilio==8.11.0
statsmodels==0.14.1
scipy<1.14
openpyxl==3.1.5
torch==2.5.1
torchvision==0.20.1
transformers==4.41.2
sentence-transformers==3.0.1